# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'PrimerApiSring'","description":"'PrimerApiSring'","contact":{"name":"Maximiliano Agustin Oliva","email":"405193@tecnicatura.frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'PrimerApiSring'"}],"paths":{"/players":{"post":{"tags":["player-controller"],"summary":"save a player in the plataform","description":"Return the player saved if the fields are correct","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"200":{"description":"Succesful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"400":{"description":"The fields are invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the plataform","description":"Return the player logged in if the credentials are correct","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credential"}}},"required":true},"responses":{"404":{"description":"The credentials are invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Succesful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/login/v2":{"post":{"tags":["login-controller"],"summary":"Login a player in the plataform","description":"Return the player logged in if the credentials are correct","operationId":"loginPlayer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialV2"}}},"required":true},"responses":{"400":{"description":"username or email already exist","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Succesful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healty of the app","description":"If the app is active response pong","operationId":"ping","responses":{"200":{"description":"Succesful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}}},"components":{"schemas":{"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userName":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"avatar":{"type":"string"},"lastLogin":{"type":"string","format":"date-time"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Credential":{"type":"object","properties":{"identity":{"oneOf":[{"$ref":"#/components/schemas/EmailIdentity"},{"$ref":"#/components/schemas/UserNameIdentity"}]},"password":{"type":"string"}}},"EmailIdentity":{"required":["email","identity_type"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"email":{"title":"Email used to login","type":"string","description":"The player email","example":"email@email.com"}}}]},"Identity":{"required":["identity_type"],"type":"object","properties":{"identity_type":{"title":"Type of identity used to login","type":"string","description":"The type of identity provided to logged in","example":"USERNAME or EMAIL","enum":["USERNAME","EMAIL"]}},"discriminator":{"propertyName":"identity_type"}},"UserNameIdentity":{"required":["identity_type","user_name"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"user_name":{"title":"Username","type":"string","description":"The player username","example":"myUserName"}}}]},"CredentialV2":{"required":["identity","password"],"type":"object","properties":{"password":{"type":"string"},"identity":{"type":"string"}}}}}}